refactor: 💡 Move message creation out from jwt.js
