+++ client/lib/connection.js
@@ -1,9 +1,9 @@
 
-const registrationHandler = ({ payload }, res) => {
+const connectionInitHandler = ({ payload }, res) => {
   console.log('Handling connection', payload)
   res.sendStatus(200)
 }
 
 module.exports = {
-  registrationHandler
+  connectionInitHandler
 }
